requirements.txt
Django==1.6.5
Markdown==2.6.2
Pillow==2.7.0
South==1.0.2
Unipath==1.0
Werkzeug==0.10.1
Whoosh==2.7.0
argparse==1.2.1
astroid==1.3.4
bbcode==1.0.21
beautifulsoup4==4.3.2
blessings==1.6
boto==2.36.0
bpython==0.14.1
chartkick==0.4.2
curtsies==0.1.19
dj-database-url==0.3.0
dj-static==0.0.6
django-annoying==0.8.2
django-blog-zinnia==0.14.1
django-boto==0.3.9
django-ckeditor==4.4.7
django-extensions==1.5.0
django-grappelli==2.5.6
django-haystack==2.3.1
django-mockups==0.4.8
django-mptt==0.6.1
django-storages==1.1.8
django-tagging==0.3.4
django-tinymce==1.5.3
django-widget-tweaks==1.3
django-xmlrpc==0.1.5
elasticsearch==1.4.0
greenlet==0.4.5
gunicorn==19.2.1
ipdb==0.8
ipython==2.4.0
logilab-common==0.63.2
ndg-httpsclient==0.3.3
psycopg2==2.6
pyOpenSSL==0.15.1
pyasn1==0.1.7
pyelasticsearch==1.2.3
pylint==1.4.1
pyparsing==2.0.3
python-dateutil==2.4.0
pytz==2014.10
requests==2.6.2
simplejson==3.6.5
six==1.9.0
sorl-thumbnail==12.2
static3==0.5.1
urllib3==1.10.2
wsgiref==0.1.2
zinnia-wysiwyg-ckeditor==1.2
